+++ test/cases/loader-option-preprocessor-handlebars-helpers-path/src/views/helpers2/italic.js
@@ -1,0 +1,2 @@
+const Handlebars = require('handlebars');
+module.exports = (options) => new Handlebars.SafeString(`<i>${options.fn(this)}</i>`);