+++ test/cases/loader-option-preprocessor-handlebars-helpers-path/src/views/helpers2/wrapper/span.js
@@ -1,0 +1,3 @@
+const Handlebars = require('handlebars');
+// test the helper with namespace `wrapper/span`
+module.exports = (options) => new Handlebars.SafeString(`<span>${options.fn(this)}</span>`);