+++ test/utils/file.js
@@ -25,6 +25,28 @@ export const readDirRecursiveSync = function (dir = './', returnAbsolutePath = t
 };
 
 /**
+ * Get directories only.
+ * @param {string} dir
+ * @param {RegExp} test Include dirs matching this RegExp.
+ * @return {[]}
+ */
+export const readDirOnlyRecursiveSync = function (dir = './', test = null) {
+  dir = path.resolve(dir);
+
+  const entries = fs.readdirSync(dir, { withFileTypes: true });
+  const folders = entries.filter((folder) => folder.isDirectory());
+  const result = [];
+
+  for (const folder of folders) {
+    const current = path.join(dir, folder.name);
+    if (!test || test.test(current)) result.push(current);
+    result.push(...readDirOnlyRecursiveSync(current, test));
+  }
+
+  return result;
+};
+
+/**
  * Copy recursive, like as `cp -R`.
  * @param {string} src The path to the thing to copy.
  * @param {string} dest The path to the new copy.
@@ -43,6 +65,13 @@ export const copyRecursiveSync = function (src, dest) {
   } else {
     fs.copyFileSync(src, dest);
   }
+};
+
+export const removeDirsSync = function (dir, test) {
+  if (dir === '/') return;
+
+  const dirs = readDirOnlyRecursiveSync(dir, test);
+  dirs.forEach((current) => fs.rmSync(current, { recursive: true, force: true }));
 };
 
 /**