+++ test/webpack.common.js
@@ -1,4 +1,7 @@
 module.exports = {
+  output: {
+    clean: true,
+  },
   plugins: [],
 
   module: {